permissions and AI Apis

+++ build/api/controllers/rolesPermissions/rolesPermissions.controller.js
@@ -6,10 +6,20 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const constants_util_1 = __importDefault(require("../../../utils/constants.util"));
 const responseHelper_util_1 = __importDefault(require("../../../utils/responseHelper.util"));
 const rolesPermissions_service_1 = __importDefault(require("../../services/rolesPermissions.service"));
+const common_util_1 = __importDefault(require("../../../utils/common.util"));
 class RolesPermissionsController {
     constructor() {
         this.createRole = async (req, res) => {
             try {
+                //   const checkPermission = await commonUtil.checkPermission('addRole', req);
+                //   if (!checkPermission)
+                //     return res
+                //       .status(constants.CODE.BAD_REQUEST)
+                //       .send(
+                //         responseHelper.get4xxResponse(
+                //           'You do not have permission to perform this operation'
+                //         )
+                //       );
                 const response = await this.rolesPermissionsService.createRole(req);
                 if (!response[0]) {
                     return res
@@ -30,6 +40,11 @@ class RolesPermissionsController {
             }
         };
         this.getAllRoles = async (req, res) => {
+            const checkPermission = await common_util_1.default.checkPermission('viewRoles', req);
+            if (!checkPermission)
+                return res
+                    .status(constants_util_1.default.CODE.BAD_REQUEST)
+                    .send(responseHelper_util_1.default.get4xxResponse('You do not have permission to perform this operation'));
             try {
                 const response = await this.rolesPermissionsService.getAllRoles(req);
                 if (!response[0]) {
@@ -92,6 +107,11 @@ class RolesPermissionsController {
         };
         this.updateRole = async (req, res) => {
             try {
+                const checkPermission = await common_util_1.default.checkPermission('editRole', req);
+                if (!checkPermission)
+                    return res
+                        .status(constants_util_1.default.CODE.BAD_REQUEST)
+                        .send(responseHelper_util_1.default.get4xxResponse('You do not have permission to perform this operation'));
                 const response = await this.rolesPermissionsService.updateRole(req);
                 if (!response[0]) {
                     return res
@@ -112,6 +132,11 @@ class RolesPermissionsController {
         };
         this.deleteRole = async (req, res) => {
             try {
+                const checkPermission = await common_util_1.default.checkPermission('deleteRole', req);
+                if (!checkPermission)
+                    return res
+                        .status(constants_util_1.default.CODE.BAD_REQUEST)
+                        .send(responseHelper_util_1.default.get4xxResponse('You do not have permission to perform this operation'));
                 const response = await this.rolesPermissionsService.deleteRole(req);
                 if (!response[0]) {
                     return res
